--- a/objects/closure.go
+++ b/objects/closure.go
@@ -7,7 +7,7 @@ import (
 // Closure represents a function closure.
 type Closure struct {
 	Fn   *CompiledFunction
-	Free []*FreeVar
+	Free []*ObjectPtr
 }
 
 // TypeName returns the name of the type.
@@ -29,7 +29,7 @@ func (o *Closure) BinaryOp(op token.Token, rhs Object) (Object, error) {
 func (o *Closure) Copy() Object {
 	return &Closure{
 		Fn:   o.Fn.Copy().(*CompiledFunction),
-		Free: append([]*FreeVar{}, o.Free...), // DO NOT Copy() of elements; these are variable pointers
+		Free: append([]*ObjectPtr{}, o.Free...), // DO NOT Copy() of elements; these are variable pointers
 	}
 }