--- a/objects/object_ptr.go
+++ b/objects/object_ptr.go
@@ -4,38 +4,38 @@ import (
 	"github.com/d5/tengo/compiler/token"
 )
 
-// FreeVar represents a free variable.
-type FreeVar struct {
+// ObjectPtr represents a free variable.
+type ObjectPtr struct {
 	Value *Object
 }
 
-func (o *FreeVar) String() string {
+func (o *ObjectPtr) String() string {
 	return "free-var"
 }
 
 // TypeName returns the name of the type.
-func (o *FreeVar) TypeName() string {
+func (o *ObjectPtr) TypeName() string {
 	return "<free-var>"
 }
 
 // BinaryOp returns another object that is the result of
 // a given binary operator and a right-hand side object.
-func (o *FreeVar) BinaryOp(op token.Token, rhs Object) (Object, error) {
+func (o *ObjectPtr) BinaryOp(op token.Token, rhs Object) (Object, error) {
 	return nil, ErrInvalidOperator
 }
 
 // Copy returns a copy of the type.
-func (o *FreeVar) Copy() Object {
+func (o *ObjectPtr) Copy() Object {
 	return o
 }
 
 // IsFalsy returns true if the value of the type is falsy.
-func (o *FreeVar) IsFalsy() bool {
+func (o *ObjectPtr) IsFalsy() bool {
 	return o.Value == nil
 }
 
 // Equals returns true if the value of the type
 // is equal to the value of another object.
-func (o *FreeVar) Equals(x Object) bool {
+func (o *ObjectPtr) Equals(x Object) bool {
 	return o == x
 }