--- a/runtime/frame.go
+++ b/runtime/frame.go
@@ -6,9 +6,8 @@ import (
 
 // Frame represents a function call frame.
 type Frame struct {
-	fn            *objects.CompiledFunction
-	freeVars      []*objects.FreeVar
-	localFreeVars map[int]*objects.FreeVar
-	ip            int
-	basePointer   int
+	fn          *objects.CompiledFunction
+	freeVars    []*objects.ObjectPtr
+	ip          int
+	basePointer int
 }