… cmd/bench/main.go
… compiler/compiler_test.go
… runtime/vm_assignment_test.go
… runtime/vm_module_test.go
… runtime/vm_test.go
… script/compiled.go
… script/script.go
… script/variable.go
… stdlib/stdlib.go