Double indirection is gone for all of the VM except free vars. For free vars, the object is stored in objects.FreeVar.Value and then used accordingly. 

OLD:
```
-------------------------------------
fibonacci(35)
-------------------------------------
Result:  9227465
Go:      55.639295ms
Parser:  50.693µs
Compile: 105.823µs
VM:      4.337952964s
-------------------------------------
fibonacci(35) (tail-call #1)
-------------------------------------
Result:  9227465
Go:      76.950689ms
Parser:  24.202µs
Compile: 71.432µs
VM:      4.44593454s
-------------------------------------
fibonacci(35) (tail-call #2)
-------------------------------------
Result:  9227465
Go:      509ns
Parser:  33.497µs
Compile: 88.749µs
VM:      18.793µs
```

NEW:
```
-------------------------------------
fibonacci(35)
-------------------------------------
Result:  9227465
Go:      57.868119ms
Parser:  48.051µs
Compile: 103.497µs
VM:      3.155774638s
-------------------------------------
fibonacci(35) (tail-call #1)
-------------------------------------
Result:  9227465
Go:      77.439483ms
Parser:  23.099µs
Compile: 32.781µs
VM:      3.256651909s
-------------------------------------
fibonacci(35) (tail-call #2)
-------------------------------------
Result:  9227465
Go:      351ns
Parser:  19.647µs
Compile: 26.19µs
VM:      19.461µs
```