@d5 2019-03-13
I think we can combine `OpGetLocalPtr` and `OpSetLocalPtr` like this:

```golang
case compiler.OpGetLocalPtr:
	localIndex := int(v.curInsts[v.ip+1])
	v.ip++

	sp := v.curFrame.basePointer + localIndex
	val := v.stack[sp]

	var freeVar *objects.FreeVar
	if obj, ok := val.(*objects.FreeVar); ok {
		freeVar = obj
	} else {
		freeVar = &objects.FreeVar{Value: &val}
		v.stack[sp] = freeVar
	}

	if v.sp >= StackSize {
		v.err = ErrStackOverflow
		return
	}

	v.stack[v.sp] = freeVar
	v.sp++
```

Also, if we do this, we can remove `Frame.localFreeVars` as well.

Why this slows down `fib35()` is still a mystery though.